// client/src/app/models/user.model.ts
export class User {
  facebookId: string;
  _id?: string;
  email?: string;
  birthday?: string;
  firstName?: string;
  lastName?: string;
  gender?: string;
  relationshipStatus?: string;

  isSoldier?: boolean;
  isStudent?: boolean;
  isRentingApartment?: boolean;

  likeSport?: boolean;
  likeTechnology?: boolean;
  likeTours?: boolean;
  likeCooking?: boolean;
  likeMusic?: boolean;
  likeArt?: boolean;
  likeFinance?: boolean;
  likePolitics?: boolean;
}